Fix templates activities, move methods to extension, clean infrastructure
