+++ DevelopGitHub/GitHubExtension/GitHubExtension.Activity.Internal.WebApi/Queries/ActivityContextQuery.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 
 using GitHubExtension.Activity.DAL;
