--- DevelopGitHub/GitHubExtension/GitHubExtension.Infrastructure.ActionFilters/InternalActivitiesFilters/CreateIssueTemplateActivityAttribute.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Web.Http.Filters;
-using GitHubExtension.Activity.DAL;
-
-namespace GitHubExtension.Infrastructure.ActionFilters.InternalActivitiesFilters
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class CreateIssueTemplateActivityAttribute : TemplateActivityAttribute
-    {
-        public override void OnActionExecuted(HttpActionExecutedContext actionExecutedContext)
-        {
-           SaveTemplateActivity(actionExecutedContext, ActivityTypeNames.CreateIssueRequestTemplate);
-        }
-    }
-}
