+++ DevelopGitHub/GitHubExtension/GitHubExtension.Infrastructure.Extensions/GitHubExtension.Infrastructure.Extensions.csproj
@@ -65,6 +65,7 @@
     <Compile Include="HttpMessageExtensions.cs" />
     <Compile Include="HttpResponseMessageExtension.cs" />
     <Compile Include="Identity\AccessTokenExtensions.cs" />
+    <Compile Include="Identity\CurrentProjectExtensions.cs" />
     <Compile Include="Owin\AuthenticationExtension.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Owin\UseOwinContextInjectorExtension.cs" />
