+++ DevelopGitHub/GitHubExtension/GitHubExtension.Templates/Controllers/TemplatesExternalController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using System.Web.Http;
+using GitHubExtension.Activity.DAL;
 using GitHubExtension.Infrastructure.ActionFilters.InternalActivitiesFilters;
 using GitHubExtension.Templates.Commands;
 using GitHubExtension.Templates.CommunicationModels;
@@ -63,7 +64,7 @@ namespace GitHubExtension.Templates.Controllers
 
         [HttpPost]
         [Route(RouteTemplatesConstants.PullRequestTemplate)]
-        [CreatePullRequestTemplateActivity]
+        [TemplateActivity(ActivityTypeName = ActivityTypeNames.CreatePullRequestTemplate)]
         public async Task<IHttpActionResult> CreatePullRequestTemplate([FromBody]CreateUpdateTemplateModel model)
         {
             var repositoryName = User.GetCurrentProjectName();
@@ -80,6 +81,7 @@ namespace GitHubExtension.Templates.Controllers
 
         [HttpPost]
         [Route(RouteTemplatesConstants.IssueTemplate)]
+        [TemplateActivity(ActivityTypeName = ActivityTypeNames.CreateIssueRequestTemplate)]
         public async Task<IHttpActionResult> CreateIssueTemplate([FromBody]CreateUpdateTemplateModel model)
         {
             var repositoryName = User.GetCurrentProjectName();
@@ -95,6 +97,7 @@ namespace GitHubExtension.Templates.Controllers
 
         [HttpPut]
         [Route(RouteTemplatesConstants.PullRequestTemplate)]
+        [TemplateActivity(ActivityTypeName = ActivityTypeNames.UpdatePullRequestTemplate)]
         public async Task<IHttpActionResult> UpdatePullRequestTemplate([FromBody]CreateUpdateTemplateModel model)
         {
             var repositoryName = User.GetCurrentProjectName();
@@ -110,6 +113,7 @@ namespace GitHubExtension.Templates.Controllers
 
         [HttpPut]
         [Route(RouteTemplatesConstants.IssueTemplate)]
+        [TemplateActivity(ActivityTypeName = ActivityTypeNames.UpdateIssueRequestTemplate)]
         public async Task<IHttpActionResult> UpdateIssueTemplate([FromBody]CreateUpdateTemplateModel model)
         {
             var repositoryName = User.GetCurrentProjectName();
