+++ DevelopGitHub/GitHubExtension/GitHubExtension.Templates/GitHubExtension.Templates.csproj
@@ -112,6 +112,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\GitHubExtension.Activity.DAL\GitHubExtension.Activity.DAL.csproj">
+      <Project>{E8B59A6B-7115-479D-AF6E-B9047499EBE9}</Project>
+      <Name>GitHubExtension.Activity.DAL</Name>
+    </ProjectReference>
     <ProjectReference Include="..\GitHubExtension.Infrastructure.ActionFilters\GitHubExtension.Infrastructure.ActionFilters.csproj">
       <Project>{A32113BC-F69A-4DBE-8790-DF1757B7D2E6}</Project>
       <Name>GitHubExtension.Infrastructure.ActionFilters</Name>
